+++ JavaScriptSPA/authPopup.js
@@ -5,7 +5,6 @@ const myMSALObj = new Msal.UserAgentApplication(msalConfig);
 function signIn() {
   myMSALObj.loginPopup(loginRequest)
     .then(loginResponse => {  
-        console.log(loginResponse);
         console.log('id_token acquired at: ' + new Date().toString());
         
         if (myMSALObj.getAccount()) {
@@ -31,9 +30,24 @@ function getTokenPopup(request) {
       // fallback to interaction when silent call fails
       return myMSALObj.acquireTokenPopup(request)
         .then(tokenResponse => {
-
-        }).catch(error => {
+          console.log('access_token acquired at: ' + new Date().toString());
+        })
+        .catch(error => {
           console.log(error);
         });
     });
+}
+
+//calls the resource API with the token
+function callApi() {
+  getTokenPopup(tokenRequest)
+      .then(tokenResponse => {
+          console.log('access_token acquired at: ' + new Date().toString());
+          try {
+            logMessage("Request made to Web API:")
+            callApiWithAccessToken(apiConfig.webApi, tokenResponse.accessToken);
+          } catch(err) {
+            console.log(err);
+          }
+      });
 }