+++ JavaScriptSPA/policies.js
@@ -1,5 +1,5 @@
 
-// add here different b2c policies you have
+// Add here different b2c policies you have
 const b2cPolicies = {
-    signUpSignIn: "b2c_1_susi", // Enter here the policy name string you created on the portal
+    signUpSignIn: "b2c_1_susi", // enter here the policy name string you created on the portal
 }